django/requirements.txt
Django
Pillow
djsingleton
django-phonenumber-field
django-cors-headers
dateutils
phonenumbers
mysqlclient
redis
celery
requests
requests_oauthlib
pytz
pywebpush
sentry-sdk
sentry-sdk[flask]
inflect
cryptography
fuzzywuzzy
channels
gunicorn
channels_redis
psycopg2-binary
daphne
python-Levenshtein
python-magic
pyjwt
mastercard-oauth1-signer
google-auth
blinker
pytest-runner<5
six==1.12.0
aiohttp
flask
git+https://github.com/Peter-Slump/python-keycloak-client.git@master
git+https://github.com/TheEnbyperor/rasa-sdk.git@master
